feat: replace rootlesskit runtime

+++ pkg/tools/mount.go
@@ -13,6 +13,8 @@ import (
 	"path/filepath"
 	"strings"
 	"syscall"
+
+	"golang.org/x/sys/unix"
 )
 
 // IsMounted checks if the given source path is mounted in the given
@@ -51,23 +53,70 @@ func Mount(src, dest string, mode uintptr) error {
 		if err = os.MkdirAll(filepath.Dir(dest), 0o755); err != nil {
 			return err
 		}
-		file, createErr := os.Create(dest)
-		if createErr != nil {
-			return createErr
+		destInfo, statErr := os.Lstat(dest)
+		if statErr == nil && destInfo.Mode()&os.ModeSymlink != 0 {
+			if removeErr := os.Remove(dest); removeErr != nil {
+				return removeErr
+			}
+			statErr = os.ErrNotExist
 		}
-		if closeErr := file.Close(); closeErr != nil {
-			return closeErr
+		if os.IsNotExist(statErr) {
+			file, createErr := os.Create(dest)
+			if createErr != nil {
+				return createErr
+			}
+			if closeErr := file.Close(); closeErr != nil {
+				return closeErr
+			}
+		} else if statErr != nil {
+			return statErr
 		}
 	}
 
 	return syscall.Mount(src, dest, "bind", mode, "")
 }
 
-// MountBind mounts bind the given source path in the given destination path.
-// It is just a wrapper around Mount, for convenience.
+// MountBind creates a recursive private bind mount.
 func MountBind(src, dest string) error {
-	return Mount(src, dest, syscall.MS_BIND|syscall.MS_REC|syscall.MS_RDONLY|syscall.MS_NOSUID|
-		syscall.MS_NOEXEC|syscall.MS_NODEV|syscall.MS_PRIVATE|syscall.MS_SLAVE)
+	if err := Mount(src, dest, syscall.MS_BIND|syscall.MS_REC); err != nil {
+		return err
+	}
+	return syscall.Mount("", dest, "", syscall.MS_PRIVATE|syscall.MS_REC, "")
+}
+
+func IsSameFile(first, second string) bool {
+	firstInfo, firstErr := os.Stat(first)
+	secondInfo, secondErr := os.Stat(second)
+	return firstErr == nil && secondErr == nil && os.SameFile(firstInfo, secondInfo)
+}
+
+// MountBindReadOnly creates a recursive bind mount and applies read-only
+// restrictions after the bind, since the initial bind ignores these flags.
+func MountBindReadOnly(src, dest string, noExec bool) error {
+	if err := MountBind(src, dest); err != nil {
+		return fmt.Errorf("bind %s to %s: %w", src, dest, err)
+	}
+
+	attributes := uint64(unix.MOUNT_ATTR_RDONLY | unix.MOUNT_ATTR_NOSUID | unix.MOUNT_ATTR_NODEV)
+	if noExec {
+		attributes |= unix.MOUNT_ATTR_NOEXEC
+	}
+	err := unix.MountSetattr(unix.AT_FDCWD, dest, unix.AT_RECURSIVE, &unix.MountAttr{Attr_set: attributes})
+	if err == nil {
+		return nil
+	}
+	if err != unix.ENOSYS && err != unix.EINVAL && err != unix.EOPNOTSUPP {
+		return fmt.Errorf("restrict bind mount %s: %w", dest, err)
+	}
+
+	flags := uintptr(syscall.MS_BIND | syscall.MS_REMOUNT | syscall.MS_RDONLY | syscall.MS_NOSUID | syscall.MS_NODEV)
+	if noExec {
+		flags |= syscall.MS_NOEXEC
+	}
+	if err = syscall.Mount("", dest, "", flags, ""); err != nil {
+		return fmt.Errorf("remount %s read-only: %w", dest, err)
+	}
+	return nil
 }
 
 // MountOverlay mounts the given lower, upper and work directories in the
@@ -87,10 +136,10 @@ func MountFuseOverlayfs(targetDir, lowerDir, upperDir, workDir string) (err erro
 }
 
 func MountTmpfs(targetDir string) (err error) {
-	c := exec.Command("mount", "-t", "tmpfs", "tmpfs", targetDir)
-	c.Stdout = os.Stdout
-	c.Stderr = os.Stderr
-	return c.Run()
+	if err = os.MkdirAll(targetDir, 0755); err != nil {
+		return err
+	}
+	return syscall.Mount("tmpfs", targetDir, "tmpfs", syscall.MS_NOSUID|syscall.MS_NODEV, "mode=0755")
 }
 
 func GetHostMounts() (mounts []string) {
